Tested and validated algorithm on Trinket. Lacks motor control functions
Script contains the function to calculate sunrise and sunset of the day. In addition to that, the whole logic of door control along time is ready and tested at different times and dates in the year with all transitions in sunrise/sunset.
Still the definition of motor functions to be defined and tested on Trinket M0.